refactor: 관리자 페이지 상품 컴포넌트 분리
