+++ src/refactoring/features/admin/components/index.ts
@@ -1,2 +1,3 @@
-export * from "./ProductManager";
-export * from "./CouponManager";
+export * from "./product/ProductSection";
+export * from "./coupon/CouponSection";
+
