+++ src/refactoring/features/cart/components/index.ts
@@ -1,5 +1,11 @@
-export * from './product/ProductList';
-export * from './cart/CartList';
-export * from './product/ProductCard';
+export * from './cart/CartSection';
 export * from './cart/CartCard';
 export * from './cart/CartCoupon';
+export * from './cart/CartTotal';
+export * from './cart/CartSelectedCoupon';
+export * from './cart/CartItemQuantityControls';
+export * from './cart/CartCouponSelect';
+
+export * from './product/ProductListSection';
+export * from './product/ProductCard';
+
