+++ src/refactoring/pages/cart/CartPage.tsx
@@ -1,8 +1,7 @@
 import { Coupon, Product } from "@/types";
 import { PageLayout } from "@/components";
 import { useCart } from "@/features/cart/hooks";
-import { CartList, ProductList } from "@/features/cart/components/";
-
+import { CartSection, ProductListSection } from "@/features/cart/components";
 interface Props {
   products: Product[];
   coupons: Coupon[];
@@ -20,8 +19,8 @@ export const CartPage = ({ products, coupons }: Props) => {
 
   return (
     <PageLayout title="장바구니">
-      <ProductList cart={cart} products={products} addToCart={addToCart} />
-      <CartList
+      <ProductListSection cart={cart} products={products} addToCart={addToCart} />
+      <CartSection
         cart={cart}
         updateQuantity={updateQuantity}
         removeFromCart={removeFromCart}
